% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{acme}
\alias{acme}
\title{Sample Data: A Simple Company with Departments}
\format{A data.tree root Node}
\usage{
data(acme)
}
\description{
acme's tree representation is accessed through its root, acme.
}
\details{
\itemize{
  \item cost, only available for leaf nodes. Cost of the project.
  \item p probability that a project will be undertaken.
}
}
\keyword{datasets}

