\name{ahp-package}
\alias{ahp-package}
\alias{ahp}
\docType{package}
\title{
Allows making decisions using the Analytic Hierarchy Process (AHP), as defined by Saaty.
}
\description{
Given a hierarchy tree of criteria, and a pairwise preference matrix for each criteria node of the tree, the package calculates the weights of different alternatives.
}
\details{
\tabular{ll}{
Package: \tab ahp\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-01-06\cr
License: \tab GPL-2\cr
}

}
\author{
Christoph Glur
Maintainer: Christoph Glur <christoph.glur@ipub.ch>
}
\references{
Saaty, T. L. (1980) The Analytic Hierarchy Process. McGraw-Hill, New York.
Saaty, T. L. (1977) A scaling methods for priorities in hierarchical structure. Journal of Mathematical
Psychology, 15, 234-281.
}

\keyword{ AHP }
\seealso{

  \code{\link[pmr:pmr-package]{pmr}}
}
\examples{

}
