% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp.R
\name{Ahp}
\alias{Ahp}
\title{Calculate the ahp weights from the symmetric AHP matrix}
\usage{
Ahp(mat)
}
\arguments{
\item{mat}{The symmetric AHP preference matrix}
}
\value{
the ahp preference weights
}
\description{
Calculate the ahp weights from the symmetric AHP matrix
}

