% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp.R
\name{AhpAdjust}
\alias{AhpAdjust}
\title{Tweak (adjust) an AHP matrix using adjustments}
\usage{
AhpAdjust(adjustments, base.matrix)
}
\arguments{
\item{adjustments}{A named vector of AHP preference shifts. For example: c(ai.CDT=0,ai.GLD=2,ai.RE=-9)}

\item{base.matrix}{The AHP preference matrix to adjust}
}
\description{
Tweak (adjust) an AHP matrix using adjustments
}
\details{
A positive shift translates our preferences in favour of the asset class, and a negative vice versa.
These shifts apply directly to our AHP preferences. So, if initially (in base.matrix) we prefer commodities over gold
by 3/1, then having +1 as shift will lead to 4/1. In that sense, it is recommended to have shift values between -9 and +9.
Larger values are cut off.
}
\examples{
#let's shift our preferences massively away from real estate, and slightly towards gold
library(SAA.CH.CHF)
AhpAdjust(c(ai.CDT=0,ai.GLD=2,ai.RE=-9), market.ai.base)
}

