% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp_alternative.R
\docType{class}
\name{AhpAlternative}
\alias{AhpAlternative}
\title{AHP alternative}
\format{An \code{\link{R6Class}} generator object}
\usage{
AhpAlternative$new()
}
\description{
An alternative in the AHP problem.
}
\keyword{ahp}

