% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp_alternative_node.R
\docType{class}
\name{AhpAlternativeNode}
\alias{AhpAlternativeNode}
\title{An alternative}
\format{An \code{\link{R6Class}} generator object}
\usage{
AhpAlternativeNode
}
\description{
An alternative
}
\seealso{
\code{\link{AhpNode}}
}
\keyword{datasets}

