% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp.R
\name{AhpMatrix}
\alias{AhpMatrix}
\title{Create the AHP pairwise preference matrix from vectors}
\usage{
AhpMatrix(preferenceCombinations)
}
\arguments{
\item{cat1}{A vector of n category names}

\item{cat2}{A vector of n category names}

\item{pref}{A vector of preferences}
}
\description{
Create the AHP pairwise preference matrix from vectors
}
\details{
Creates and fills a AHP conform matrix given the category vectors cat1, cat2, and the preference vector pref.
}
\examples{
categories <- combn(c('red','blue','green'),2)
cat1 <- categories[1,]
cat2 <- categories[2,]
pref <- c(1,5,4)
AhpMatrix(cat1, cat2, pref)
}
\seealso{
\code{\link{ahp}}
}

