% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ahp_node.R
\docType{class}
\name{AhpNode}
\alias{AhpNode}
\title{A node in the Ahp Tree}
\format{An \code{\link{R6Class}} generator object}
\usage{
AhpNode
}
\description{
A node in the Ahp Tree
}
\seealso{
\code{\link{Node}}
}
\keyword{datasets}

