% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alternative.R
\docType{class}
\name{Alternative}
\alias{Alternative}
\title{AHP alternative}
\format{An \code{\link{R6Class}} generator object}
\usage{
Alternative$new()
}
\description{
An alternative in the AHP problem.
}
\keyword{ahp}

