% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alternative_node.R
\docType{class}
\name{AlternativeNode}
\alias{AlternativeNode}
\title{An alternative}
\format{An \code{\link{R6Class}} generator object}
\usage{
AlternativeNode
}
\description{
An alternative
}
\seealso{
\code{\link{AhpNode}}
}
\keyword{datasets}

