% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/alternative.R
\name{AlternativesList}
\alias{AlternativesList}
\title{Constructor for a list of alternatives}
\usage{
AlternativesList(...)
}
\description{
Constructor for a list of alternatives
}
\seealso{
\code{\link{AhpAlternative}}
}

