% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{AreNamesUnique}
\alias{AreNamesUnique}
\title{Test whether all node names are unique.}
\usage{
AreNamesUnique(node)
}
\arguments{
\item{node}{The root \code{Node} of the \code{data.tree} structure to test}
}
\value{
\code{TRUE} if all \code{Node$name == TRUE} for all nodes in the tree
}
\description{
This can be useful for some conversions.
}
\examples{
data(acme)
AreNamesUnique(acme)
acme$name <- "IT"
AreNamesUnique(acme)
}
\seealso{
as.igraph.Node
}

