% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dendrogram.R
\name{as.dendrogram.Node}
\alias{as.dendrogram.Node}
\title{Convert a Node to a dendrogram}
\usage{
\method{as.dendrogram}{Node}(object, heightAttribute = Height, ...)
}
\arguments{
\item{object}{The Node to convert}

\item{heightAttribute}{The attribute (field name or function) storing the height}

\item{...}{Additional parameters}
}
\value{
An object of class dendrogram
}
\description{
Convert a Node to a dendrogram
}
\details{
Convert a Node to a dendrogram
}
\examples{
data(acme)
acmed <- as.dendrogram(acme)
plot(acmed, center = TRUE)

#you can take an attribute for the height:
acme$Do( function(x) x$height <- (10 - x$level))
acmed <- as.dendrogram(acme, heightAttribute = "height")
plot(acmed, center = TRUE)

#or directly a function
acmed <- as.dendrogram(acme, heightAttribute = function(x) 10 - x$level)
plot(acmed)
}
\seealso{
Other Conversions from Node: \code{\link{ToListExplicit}};
  \code{\link{ToListSimple}}; \code{\link{ToNewick}};
  \code{\link{as.list.Node}}
}

