% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{as.list.Node}
\alias{as.list.Node}
\title{Convert a Node to a list}
\usage{
\method{as.list}{Node}(x, mode = c("simple", "explicit"), unname = FALSE,
  nameName = ifelse(unname, "name", ""), childrenName = "children",
  rootName = "", ...)
}
\arguments{
\item{x}{The Node to convert}

\item{mode}{How the list is structured. "simple" (the default) will add children directly as nested lists.
"explicit" puts children in a separate nested list called \code{childrenName}}

\item{unname}{If TRUE, then the nested children list will not have named arguments. This
can be useful e.g. in the context of conversion to JSON, if you prefer the children to be
an array rather than named objects.}

\item{nameName}{The name that should be given to the name element}

\item{childrenName}{The name that should be given to the children nested list}

\item{...}{Additional parameters}

\item{nodeName}{The name of the node. If given, this overrides the name of the Node}
}
\description{
Convert a Node to a list
}
\details{
Convert a Node to a list
}

