% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{as.Node.list}
\alias{as.Node.list}
\title{Converts a list to a Node}
\usage{
as.Node.list(x, ...)
}
\arguments{
\item{x}{The list to be converted.}

\item{...}{Any other argument to be passed to generic sub implementations}
}
\description{
Converts a list to a Node
}
\details{
x should contained named elements, where the name will be converted to the Node's public attribute
name, and the value to its value.
x should not contain any element whose name is in NODE_RESERVED_NAMES_CONST, except the ones mentioned below.
x has to contain an element called name, which must be convertible to a character string.
x can contain an element called children, which should be of class list. This will be converted to the Node's
children.
}

