% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_ape.R
\name{as.Node.phylo}
\alias{as.Node.phylo}
\title{Converts a phylo from the ape package to a Node}
\usage{
\method{as.Node}{phylo}(p, heightName = "height", replaceUnderscores = TRUE)
}
\arguments{
\item{p}{The phylo object}

\item{heightName}{If the phylo contains edge lengths, then they will be stored in
an attribute according to this parameter (the default is "edgeLength")}

\item{replaceUnderscores}{if TRUE (the default), then underscores in names are replaced with spaces}
}
\description{
Converts a phylo from the ape package to a Node
}

