% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{as.Node}
\alias{as.Node}
\title{Convert an object to a Node}
\usage{
as.Node(x, ...)
}
\arguments{
\item{x}{The object to be converted}

\item{...}{Additional arguments}
}
\description{
Convert an object to a Node
}
\details{
Convert an object to a Node
}

