% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{as.phylo.Node}
\alias{as.phylo.Node}
\title{Convert a node to a phylo object from the ape package.}
\usage{
\method{as.phylo}{Node}(x, heightAttributeName = "Height", ...)
}
\arguments{
\item{x}{The \code{Node} to convert}

\item{heightAttributeName}{To use custom heights}

\item{...}{any other argument}
}
\description{
Convert a node to a phylo object from the ape package.
}

