% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ball}
\alias{ball}
\title{Is Jack playing ball today?}
\format{data.frame}
\usage{
data(ball)
}
\description{
Ball contains daily weather conditions, and whether hypothetical Jack, a tennis player,
played ball on a given day. We can use this data for classification trees, to check whether
we can predict if Jack will play ball on a given day, given the current weather conditions.
ball contains five factor attributes:
}
\details{
\itemize{
  \item outlook the weather outlook
  \item temperature The temperature of the day (Hot, Mild, Cool, Cold)
  \item humidity The humidity (High, Normal, Low)
  \item wind The wind on the day (Weak, Strong)
  \item play Whether Jack plays tennis on the given days (Yes, No)
}
}
\keyword{datasets}

