% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{Climb}
\alias{Climb}
\title{Find a \code{Node} by its path}
\usage{
Climb(node, ...)
}
\arguments{
\item{node}{The root node}

\item{...}{the names of the nodes in the path}
}
\value{
the \code{Node} having path \code{...}, or \code{NULL} if such a path does not exist
}
\description{
\code{Climb} returns the \code{Node} at path \code{...}. The path is relative to the \code{Node} on which this method is called. Each argument provided corresponds to an
element in the path, specified by the \code{Node}'s name.
}
\examples{
data(acme)
acme$Climb('IT', 'Outsource')$name
acme$Climb('IT')$Climb('Outsource')$name
}
\seealso{
\code{\link{Node}}
}

