% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{Clone}
\alias{Clone}
\title{Clone a tree (creates a deep copy)}
\usage{
Clone(node)
}
\arguments{
\item{node}{the root node of the tree or sub-tree to clone}
}
\value{
the clone of the tree
}
\description{
The method also clones object attributes (such as the formatters).
}
\examples{
data(acme)
acmeClone <- Clone(acme)
acmeClone$name <- "New Acme"
# acmeClone does not point to the same reference object anymore:
acme$name
}
\seealso{
SetFormat
}

