% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{CreateDummyTree}
\alias{CreateDummyTree}
\title{Create a tree for demo and testing}
\usage{
CreateDummyTree(levels = 5, children = 3, parent = Node$new("1"))
}
\arguments{
\item{levels}{the number of levels}

\item{children}{the number of children per node}

\item{parent}{the parent node (for recursion)}
}
\description{
Create a tree for demo and testing
}

