% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_conversion_party.R
\name{CreateNodeFromParty}
\alias{CreateNodeFromParty}
\title{Convert a a \code{SplitNode} from the party package to a \code{data.tree} structure.}
\usage{
CreateNodeFromParty(splittingNode)
}
\arguments{
\item{splittingNode}{a SplittingNode object}
}
\description{
Convert a a \code{SplitNode} from the party package to a \code{data.tree} structure.
}
\examples{
library(party)
airq <- subset(airquality, !is.na(Ozone))
airct <- ctree(Ozone ~ ., data = airq, 
               controls = ctree_control(maxsurrogate = 3))
               
tree <- CreateNodeFromParty(airct@tree)
tree

print(tree, 
      psplit = function(x) as.character(x$psplit), 
      criterion = function(x) round(x$criterion$maxcriterion, 3),
      statistic = function(x) round(max(x$criterion$statistic))
      )

tree$FindNode(6)$path


}

