% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{CreateRandomTree}
\alias{CreateRandomTree}
\title{Create a tree for demo and testing}
\usage{
CreateRandomTree(nodes = 100, root = Node$new("1"), id = 1)
}
\arguments{
\item{nodes}{The number of nodes to create}

\item{root}{the previous node (for recursion, typically use default value)}

\item{id}{The id (for recursion)}
}
\description{
Create a tree for demo and testing
}

