% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{CreateRegularTree}
\alias{CreateRegularTree}
\title{Create a tree for demo and testing}
\usage{
CreateRegularTree(height = 5, branchingFactor = 3,
  parent = Node$new("1"))
}
\arguments{
\item{height}{the number of levels}

\item{branchingFactor}{the number of children per node}

\item{parent}{the parent node (for recursion)}
}
\description{
Create a tree for demo and testing
}
