\name{data.tree-package}
\alias{data.tree-package}
\alias{data.tree}
\docType{package}
\title{
Construct hierarchical tress from data
}
\description{
Hierarchical data is ubiquitous. However, no general-use tree data structure is available in R, similar to, say, data.frame for tabular data. This package allows building trees of hierarchical data, traverse the trees and run operations on nodes. The package also contains utility functions to sort, to print, and to construct a tree from a data.frame.
}
\details{
\tabular{ll}{
Package: \tab data.tree\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2015-01-06\cr
License: \tab GPL-2\cr
}

}
\author{
Christoph Glur
Maintainer: Christoph Glur <christoph.glur@ipub.ch>
}
\references{
}

\keyword{ tree }
\seealso{


}
\examples{

}
