% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.tree.R
\docType{package}
\name{data.tree}
\alias{data.tree}
\alias{data.tree-package}
\title{data.tree hierarchical data for R}
\description{
Hierarchical data is ubiquitous. However, no general-use tree data structure is available in R.
Where tabular data has \code{data.frame}, hierarchical data is often modeled in lists of list or similar ugly constructions.
The \code{data.tree} package allows building trees of hierarchical
data. It allows traversing the trees and run operations on nodes or collect data, which is very handy for
the rapid prototyping of algorithms. The package also contains
utility functions to sort, to print, and to construct a tree from a \code{data.frame} and vice versa.
}
\seealso{
\code{\link{Node}}

For more details see the \code{data.tree} vignette: \code{vignette("data.tree")}
}

