% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_methods.R
\name{Distance}
\alias{Distance}
\title{Find the distance between two nodes of the same tree}
\usage{
Distance(node1, node2)
}
\arguments{
\item{node1}{the first node in the tree}

\item{node2}{the second node in the same tree}
}
\description{
The distance is measured as the number of edges that
need to be traversed to reach node2 when starting 
from node1.
}
\examples{
data(acme)
Distance(FindNode(acme, "Outsource"), FindNode(acme, "Research"))

}

