% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/node.R
\name{Find}
\alias{Find}
\title{Finds a node having path \code{...}.}
\usage{
Find(...)
}
\arguments{
\item{...}{the names of the nodes in the path}
}
\value{
the Node having path \code{...}, or NULL if such a path does not exist
}
\description{
The path is relative to the Node on which this method is called. Each argument provided corresponds to an
element in the path.
}
\examples{
data(acme)
acme$Find('IT', 'Outsource')$name
#This is equivalent to:
acme$Find('IT')$Find('Outsource')$name
acme$Find('X', 'Y', 'Z')
}
\seealso{
\code{\link{Node}}
}
\keyword{internal}

