% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{FormatPercent}
\alias{FormatPercent}
\title{Format a Number as a Percentage}
\usage{
FormatPercent(x, digits = 2, format = "f", ...)
}
\arguments{
\item{x}{A number}

\item{digits}{The number of digits to print}

\item{format}{The format to use}

\item{...}{Any other argument passed to formatC}
}
\value{
A string corresponding to x, suitable for printing
}
\description{
This utility method can be used as a format function when converting trees to a \code{data.frame}
}
\seealso{
formatC
}

