% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{FromDataFrameTaxonomy}
\alias{FromDataFrameTaxonomy}
\title{Convert a taxonomy into a Node.}
\usage{
FromDataFrameTaxonomy(x)
}
\arguments{
\item{x}{The taxonomy data.frame to convert.}
}
\description{
To be a taxonomy, a data.frame must fulfil the following requirements:
1. It must contain as many rows as there are nodes, excluding the root
2. Its first column must be called *children*, and contain the name of each node, whereby each name must be unique within a node's level
3. Its second column must be called *parents*, and contain the name of the parent of each node
4. Its third column must be called *level*, and contain the level of the node
5. The rows must be ordered by level
}
\examples{
data(acme)
x <- ToDataFrameTaxonomy(acme, "p", "cost")
xN <- FromDataFrameTaxonomy(x)
print(xN, "p", "cost")
}
\seealso{
Other Conversions to Node: \code{\link{as.Node.data.frame}};
  \code{\link{as.Node.dendrogram}};
  \code{\link{as.Node.list}}

Other data.frame Conversions: \code{\link{ToDataFrameTable}};
  \code{\link{ToDataFrameTaxonomy}};
  \code{\link{ToDataFrameTree}};
  \code{\link{as.Node.data.frame}};
  \code{\link{as.data.frame.Node}}
}

