% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods_traversal.R
\name{Get}
\alias{Get}
\title{Traverse a Tree and Collect Values}
\usage{
Get(nodes, attribute, ..., format = NULL, inheritFromAncestors = FALSE,
  simplify = TRUE)
}
\arguments{
\item{nodes}{The nodes on which to perform the Get (e.g. obtained via \code{\link{Traverse}}}

\item{attribute}{determines what is collected. The \code{attribute} can be
\itemize{
  \item a.) the name of a \bold{field} of each \code{Node} in the tree
  \item b.) the name of a \bold{method} of each \code{Node} in the tree
  \item c.) a \bold{function}, whose first argument must be a \code{Node}
 }}

\item{...}{in case the \code{attribute} is a function or a method, the ellipsis is passed to it as additional arguments.}

\item{format}{can be a function that transforms the collected values, e.g. for printing}

\item{inheritFromAncestors}{if \code{TRUE}, then the path above a \code{Node} is searched to get the \code{attribute} in case it is NULL.}

\item{simplify}{logical or character string; should the result be
    simplified to a vector, matrix or higher dimensional array if
    possible?  For \code{sapply} it must be named and not abbreviated.
    The default value, \code{TRUE}, returns a vector or matrix if appropriate,
    whereas if \code{simplify = "array"} the result may be an
    \code{\link{array}} of \dQuote{rank}
    (\eqn{=}\code{length(dim(.))}) one higher than the result
    of \code{FUN(X[[i]])}.}
}
\value{
a vector containing the \code{atrributes} collected during traversal, in traversal order. \code{NULL} is converted
to NA, such that \code{length(Node$Get) == Node$totalCount}
}
\description{
The \code{Get} method is one of the most important ones of the \code{data.tree} package. It lets you traverse a tree
and collect values along the way. Alternatively, you can call a method or a function on each \code{\link{Node}}.
}
\examples{
data(acme)
acme$Get("level")
acme$Get("totalCount")

calculateAggregateChildCost <- function(node, fun) {
 if (node$isLeaf) return(node$cost)
 fun(sapply(node$children, function(x) x$averageCost))
}

myFormat <- function(x) {
 format(x, nsmall=2, scientific = FALSE)
}

acme$Get(calculateAggregateChildCost,
        mean,
        traversal = "post-order",
        format = myFormat)
}
\seealso{
\code{\link{Node}}

\code{\link{Set}}
}

