% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{GetPhyloLabel}
\alias{GetPhyloLabel}
\title{Get a Phylo Label for a single Node}
\usage{
GetPhyloLabel(x, labelFun, type = c("node", "edge"), ...)
}
\arguments{
\item{x}{The Node}

\item{labelFun}{The function that compiles the label}

\item{type}{either "node" (the default) or "edge"}

\item{...}{any argument to be passed to the labelFun}
}
\value{
a string that can be used as a Phylo Node label
}
\description{
Get a Phylo Label for a single Node
}

