% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{GetPhyloLabels}
\alias{GetPhyloLabels}
\title{Get a Phylo Labels for a single Node}
\usage{
GetPhyloLabels(x, labelFun, ...)
}
\arguments{
\item{x}{The root Node}

\item{labelFun}{The function that compiles the label}

\item{...}{any argument to be passed to the labelFun}
}
\value{
a string vector that can be used as a Phylo Node label
}
\description{
Get a Phylo Labels for a single Node
}

