% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{GetPhyloNodeNr}
\alias{GetPhyloNodeNr}
\title{Gets the node nr in phylo context}
\usage{
GetPhyloNodeNr(x)
}
\arguments{
\item{x}{The Node}
}
\value{
an integer representing the node
}
\description{
Gets the node nr in phylo context
}

