% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{GetPhyloNr}
\alias{GetPhyloNr}
\title{Gets the node nr in phylo context}
\usage{
GetPhyloNr(x, type = c("node", "edge"))
}
\arguments{
\item{x}{The Node}
}
\value{
an integer representing the node
}
\description{
Gets the node nr in phylo context
}

