% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{Height}
\alias{Height}
\title{Calculates the height of a \code{Node} given the hight of the root,
  assuming that nodes are equally distributed. Useful for easy printing
  as dendrogram.}
\usage{
Height(node, rootHeight = 100)
}
\arguments{
\item{node}{The node}

\item{rootHeight}{The height of the root}
}
\description{
Calculates the height of a \code{Node} given the hight of the root,
  assuming that nodes are equally distributed. Useful for easy printing
  as dendrogram.
}

