% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_actives.R
\name{isLeaf}
\alias{isLeaf}
\title{Check if a \code{Node} is a leaf}
\usage{
isLeaf(node)
}
\arguments{
\item{node}{The Node to test.}
}
\value{
TRUE if the Node is a leaf, FALSE otherwise
}
\description{
Check if a \code{Node} is a leaf
}
