% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_actives.R
\name{isNotLeaf}
\alias{isNotLeaf}
\title{Checks if a node is not a leaf}
\usage{
isNotLeaf(node)
}
\arguments{
\item{node}{The Node to test.}
}
\value{
FALSE if the Node is a leaf, TRUE otherwise
}
\description{
Checks if a node is not a leaf
}

