% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_actives.R
\name{isNotRoot}
\alias{isNotRoot}
\title{Checks if a node is not a root}
\usage{
isNotRoot(node)
}
\arguments{
\item{node}{The Node to test.}
}
\value{
FALSE if the Node is the root, TRUE otherwise
}
\description{
Checks if a node is not a root
}

