% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_actives.R
\name{isRoot}
\alias{isRoot}
\title{Check if a \code{Node} is the root}
\usage{
isRoot(node)
}
\arguments{
\item{node}{The Node to test.}
}
\value{
TRUE if the Node is the root, FALSE otherwise
}
\description{
Check if a \code{Node} is the root
}

