% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{Sample Data: Data Used by the ID3 Vignette}
\format{data.frame}
\usage{
data(mushroom)
}
\description{
mushroom contains attributes of mushrooms. We can use this data to predict a
mushroom's toxicity based on its attributes.
The attributes available in the data set are:
}
\details{
\itemize{
  \item color the weather outlook
  \item size The temperature of the day (Hot, Mild, Cool, Cold)
  \item points The humidity (High, Normal, Low)
  \item edibility The wind on the day (Weak, Strong)
}
}
\keyword{datasets}

