% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/util.R
\name{PrintFixedDecimal}
\alias{PrintFixedDecimal}
\title{Format a number as a decimal with a fixed number of decimals.}
\usage{
PrintFixedDecimal(x, digits = 3)
}
\arguments{
\item{x}{a numeric scalar or vector}

\item{digits}{the number of digits to print after the decimal point}
}
\value{
A string corresponding to x, suitable for printing
}
\description{
Format a number as a decimal with a fixed number of decimals.
}

