% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_conversion_igraph.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.igraph}
\title{Register the generic 'as.igraph' from package 'igraph'}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{igraph}{\code{\link[igraph]{as.igraph}}}
}}

