% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_doc.R
\name{Revert}
\alias{Revert}
\title{Reverts the sort order of a \code{Node}'s children.}
\usage{
Revert(recursive = TRUE)
}
\arguments{
\item{recursive}{If \code{TRUE}, then revert is called recursively on
all children.


#'}
}
\description{
Reverts the sort order of a \code{Node}'s children.
}
\seealso{
\code{\link{Node}}
}
\keyword{internal}

