% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{SetFormat}
\alias{SetFormat}
\title{Set a formatter function on a specific node}
\usage{
SetFormat(node, name, formatFun)
}
\arguments{
\item{node}{The node on which to set the formatter}

\item{name}{The attribute name for which to set the formatter}

\item{formatFun}{The formatter, i.e. a function taking a value as an input, and formatting
returning the formatted value}
}
\description{
Set a formatter function on a specific node
}

