% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{SetGraphStyle}
\alias{SetGraphStyle}
\title{Call this on the root \code{\link{Node}} to set style attributes
on the graph}
\usage{
SetGraphStyle(root, ...)
}
\arguments{
\item{root}{The root node}

\item{...}{Any style attribute}
}
\description{
Call this on the root \code{\link{Node}} to set style attributes
on the graph
}
\examples{
data(acme)
SetGraphStyle(acme, rankdir = "LR", layout = "twopi")
ToGraphViz(acme)
}

