% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{ToDataFrameTable}
\alias{ToDataFrameTable}
\title{Converts the tree to tabular form, keeping only the leafs.}
\usage{
ToDataFrameTable(node, ..., pruneFun = NULL, filterFun = NULL)
}
\arguments{
\item{node}{The node to convert}

\item{...}{character strings representing the attributes to be returned. Note that these may not be
available on a leaf, in which case the algorithm inherits the values from its ancestors.}

\item{pruneFun}{a function pruning the tree to be converted}

\item{filterFun}{a function filtering nodes to be converted}
}
\description{
This method is especially useful when you need to apply a specific function that is only available
for \code{\link{data.frame}}.
}

