% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{ToDataFrameTaxonomy}
\alias{ToDataFrameTaxonomy}
\title{Convert a \code{Node} to a data.frame of the type taxonomy.}
\usage{
ToDataFrameTaxonomy(node, ..., pruneFun = NULL,
  inheritFromAncestors = FALSE)
}
\arguments{
\item{node}{the root Node of the tree or sub-tree to convert}

\item{...}{the attributes to add as columns to the data frame}

\item{pruneFun}{a function allowing to prevent some sub-trees to be converted}

\item{inheritFromAncestors}{if TRUE, then attributes are inherited from the closest ancestor if not available in a node itself}
}
\description{
This is useful e.g. when converting to igraph.
}
\examples{
data(acme)
ToDataFrameTaxonomy(acme, "cost", "p")
}
\seealso{
Other Conversions from Node: \code{\link{ToDataFrameTable}};
  \code{\link{ToDataFrameTree}};
  \code{\link{ToListExplicit}}; \code{\link{ToListSimple}};
  \code{\link{ToNewick}}; \code{\link{as.data.frame.Node}};
  \code{\link{as.dendrogram.Node}};
  \code{\link{as.list.Node}}

Other data.frame Conversions: \code{\link{FromDataFrameTaxonomy}};
  \code{\link{ToDataFrameTable}};
  \code{\link{ToDataFrameTree}};
  \code{\link{as.Node.data.frame}};
  \code{\link{as.data.frame.Node}}
}

