% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{ToDataFrameTree}
\alias{ToDataFrameTree}
\title{Convert a \code{\link{Node}} to a \code{data.frame}}
\usage{
ToDataFrameTree(node, ..., pruneFun = NULL)
}
\arguments{
\item{node}{The node to convert}

\item{...}{the attributes to be added as columns of the data.frame. There are various
options:
\itemize{
 \item a string corresponding to the name of a node attribute
 \item the result of the \code{Node$Get} method
}
If a specific Node does not contain the attribute, the data.frame will contain NA.}

\item{pruneFun}{a function pruning the nodes to be converted}
}
\description{
Convert a \code{\link{Node}} to a \code{data.frame}
}
\examples{
data(acme)
ToDataFrameTree(acme, "cost", "p")

#use the pruneFun:
acme$Do(function(x) x$totalCost <- Aggregate(x, "cost", sum), traversal = "post-order")
ToDataFrameTree(acme, "totalCost", pruneFun = function(x) x$totalCost > 300000)
}
\seealso{
Other Conversions from Node: \code{\link{ToDataFrameTable}};
  \code{\link{ToDataFrameTaxonomy}};
  \code{\link{ToListExplicit}}; \code{\link{ToListSimple}};
  \code{\link{ToNewick}}; \code{\link{as.data.frame.Node}};
  \code{\link{as.dendrogram.Node}};
  \code{\link{as.list.Node}}

Other data.frame Conversions: \code{\link{FromDataFrameTaxonomy}};
  \code{\link{ToDataFrameTable}};
  \code{\link{ToDataFrameTaxonomy}};
  \code{\link{as.Node.data.frame}};
  \code{\link{as.data.frame.Node}}
}

