% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_doc.R
\name{ToList}
\alias{ToList}
\title{Convert a \code{\link{Node}} object to a nested \code{list}}
\usage{
ToList(unname = FALSE, nameName = "name", childrenName = "children", ...)
}
\arguments{
\item{unname}{If TRUE, then the nested children list will not have named arguments. This
can be useful e.g. in the context of conversion to JSON, if you prefer the children to be
an array rather than named objects.}

\item{nameName}{The name that should be given to the name element}

\item{childrenName}{The name that should be given to the children nested list}

\item{...}{Additional parameters}

\item{x}{The Node to convert}
}
\description{
Convert a \code{\link{Node}} object to a nested \code{list}
}
\seealso{
\code{\link{Node}}, \code{\link{as.list.Node}}
}
\keyword{internal}

