% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{ToListExplicit}
\alias{ToListExplicit}
\title{Export a Node to a list of lists where children are explicit, meaning there is a nested list called 'children', containing
lists of child nodes.}
\usage{
ToListExplicit(node, ..., unname = FALSE, nameName = ifelse(unname, "name",
  ""), childrenName = "children")
}
\arguments{
\item{node}{The root node to export}

\item{...}{The attributes to export}

\item{unname}{If TRUE, then children are not named in the explicit children list}

\item{nameName}{The name of the name attribute}

\item{childrenName}{the name of the children list}
}
\description{
Export a Node to a list of lists where children are explicit, meaning there is a nested list called 'children', containing
lists of child nodes.
}
\examples{
data(acme)
str(ToListExplicit(acme, "cost"))
str(ToListExplicit(acme, "cost", unname = TRUE))
str(ToListExplicit(acme, "cost", unname = TRUE, nameName = "id", childrenName = "descendants"))
}
\seealso{
Other Conversions from Node: \code{\link{ToListSimple}};
  \code{\link{ToNewick}}; \code{\link{as.dendrogram.Node}};
  \code{\link{as.list.Node}}

Other List Conversions: \code{\link{ToListSimple}};
  \code{\link{as.Node.list}}; \code{\link{as.list.Node}}
}

