% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{ToListSimple}
\alias{ToListSimple}
\title{Exports a tree to a list of lists in simple format, meaning that children are put directly into a nested list.}
\usage{
ToListSimple(node, ..., nameName = "name")
}
\arguments{
\item{node}{The root node to export}

\item{...}{the fields to export, as character strings}

\item{nameName}{the name of the name attribute}
}
\description{
Exports a tree to a list of lists in simple format, meaning that children are put directly into a nested list.
}
\examples{
data(acme)
alol <- ToListSimple(acme, "p", "cost")
str(alol)
}
\seealso{
Other Conversions from Node: \code{\link{ToListExplicit}};
  \code{\link{ToNewick}}; \code{\link{as.dendrogram.Node}};
  \code{\link{as.list.Node}}

Other List Conversions: \code{\link{ToListExplicit}};
  \code{\link{as.Node.list}}; \code{\link{as.list.Node}}
}

