% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{as.data.frame.Node}
\alias{as.data.frame.Node}
\title{Convert a \code{\link{Node}} to a \code{data.frame}}
\usage{
\method{as.data.frame}{Node}(x, row.names = NULL, optional = FALSE, ...,
  traversal = c("pre-order", "post-order", "in-order", "level", "ancestor"),
  pruneFun = NULL, filterFun = NULL, inheritFromAncestors = FALSE)
}
\arguments{
\item{x}{The root node to convert to a data.frame}

\item{row.names}{\code{NULL} or a character vector giving the row names for the data frame.
Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and converting column names
(to syntactic names: see make.names) is optional.}

\item{...}{the attributes to be added as columns of the data.frame. There are various
options:
\itemize{
 \item a string corresponding to the name of a node attribute
 \item the result of the \code{Node$Get} method
}
If a specific Node does not contain the attribute, \code{NA} is added to the data.frame.}

\item{traversal}{any of 'pre-order' (the default), 'post-order', 'in-order', 'level', or 'ancestor'}

\item{pruneFun}{a function which, if TRUE is returned, causes the subtree of a Node to be pruned.}

\item{filterFun}{a function which filters the Nodes added to the \code{data.frame}. The function must
take a \code{Node} as an input, and it must return \code{TRUE} or \code{FALSE}, depending on whether the}

\item{inheritFromAncestors}{if TRUE, then any attribute specified in \code{...} is fetched from a Node, or from any
of its ancestors
\code{Node} and its subtree should be displayed.}
}
\description{
Convert a \code{\link{Node}} to a \code{data.frame}
}

