% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dendrogram.R
\name{as.dendrogram.Node}
\alias{as.dendrogram.Node}
\title{Convert a Node to a dendrogram}
\usage{
\method{as.dendrogram}{Node}(object, heightAttribute = Height, ...)
}
\arguments{
\item{object}{The Node to convert}

\item{heightAttribute}{The attribute (field name or function) storing the height}

\item{...}{Additional parameters}
}
\description{
Convert a Node to a dendrogram
}
\details{
Convert a Node to a dendrogram
}

