% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\name{as.list.Node}
\alias{as.list.Node}
\title{Convert a Node to a list}
\usage{
\method{as.list}{Node}(x, unname = FALSE, nameName = "name",
  childrenName = "children", ...)
}
\arguments{
\item{x}{The Node to convert}

\item{unname}{If TRUE, then the nested children list will not have named arguments. This
can be useful e.g. in the context of conversion to JSON, if you prefer the children to be
an array rather than named objects.}

\item{nameName}{The name that should be given to the name element}

\item{childrenName}{The name that should be given to the children nested list}

\item{...}{Additional parameters}
}
\description{
Convert a Node to a list
}
\details{
Convert a Node to a list
}

