% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{as.Node.data.frame}
\alias{as.Node.data.frame}
\title{Convert a data.frame to a data.tree}
\usage{
\method{as.Node}{data.frame}(x, ..., pathName = "pathString",
  pathDelimiter = "/", colLevels = NULL, na.rm = FALSE)
}
\arguments{
\item{x}{The data.frame}

\item{...}{Any other argument}

\item{pathName}{The name of the column in x containing the path of the row}

\item{pathDelimiter}{The delimiter used}

\item{colLevels}{Nested vector of column names, determining on what node levels the attributes are written to.}

\item{na.rm}{If \code{TRUE}, then NA's are treated as NULL and values will not be set on nodes}
}
\description{
Convert a data.frame to a data.tree
}
\details{
x should be of class x
}
\seealso{
Other Conversions to Node: \code{\link{as.Node.dendrogram}};
  \code{\link{as.Node.list}}
}

