% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{as.Node.dendrogram}
\alias{as.Node.dendrogram}
\title{Convert a \code{\link{dendrogram}} to a data.tree \code{Node}}
\usage{
\method{as.Node}{dendrogram}(x, name = "root", ...)
}
\arguments{
\item{x}{The dendrogram}

\item{...}{Additional parameters}
}
\value{
The root \code{Node} of a \code{data.tree}
}
\description{
Convert a \code{\link{dendrogram}} to a data.tree \code{Node}
}

