% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{as.Node}
\alias{as.Node}
\title{Convert an object to a \code{data.tree} data structure}
\usage{
as.Node(x, ...)
}
\arguments{
\item{x}{The object to be converted}

\item{...}{Additional arguments}
}
\description{
Convert an object to a \code{data.tree} data structure
}
\seealso{
Other as.Node: 
\code{\link{as.Node.data.frame}()},
\code{\link{as.Node.dendrogram}()},
\code{\link{as.Node.list}()},
\code{\link{as.Node.phylo}()},
\code{\link{as.Node.rpart}()}
}
\concept{as.Node}
