% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_ape.R
\name{as.phylo.Node}
\alias{as.phylo.Node}
\title{Convert a node to a phylo object from the ape package.
This requires the ape package.}
\usage{
as.phylo.Node(x, heightAttribute = Height, ...)
}
\arguments{
\item{x}{The \code{Node} to convert}

\item{...}{any other argument}

\item{heightAttributeName}{To use custom heights}
}
\description{
Convert a node to a phylo object from the ape package.
This requires the ape package.
}

