% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_conversion_ape.R
\name{as.phylo.Node}
\alias{as.phylo.Node}
\title{Convert a \code{Node} to a phylo object from the ape package.}
\usage{
as.phylo.Node(x, heightAttribute = DefaultPlotHeight, ...)
}
\arguments{
\item{x}{The root \code{Node} of the tree or sub-tree to be converted}

\item{heightAttribute}{The attribute (field name or function) storing the height}

\item{...}{any other argument}
}
\description{
This method requires the ape package to be installed and loaded.
}
\examples{
library(ape)
data(acme)
acmephylo <- as.phylo(acme)
#plot(acmephylo)


}
\seealso{
Other ape phylo conversions: \code{\link{GetPhyloNr}},
  \code{\link{as.Node.phylo}}
}
\concept{ape phylo conversions}
