% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_actives.R
\name{averageBranchingFactor}
\alias{averageBranchingFactor}
\title{Calculate the average number of branches each non-leaf has}
\usage{
averageBranchingFactor(node)
}
\arguments{
\item{node}{The node}
}
\description{
Calculate the average number of branches each non-leaf has
}
