% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_methods.R
\name{Climb}
\alias{Climb}
\title{Climb a tree from parent to children, by provided criteria.}
\usage{
#node$Climb(...)
Climb(node, ...)
}
\arguments{
\item{node}{The root \code{\link{Node}} of the tree or subtree to climb}

\item{...}{an attribute-value pairlist to be searched. For brevity, you can also provide a character vector to search for names.}
}
\value{
the \code{Node} having path \code{...}, or \code{NULL} if such a path does not exist
}
\description{
This method lets you climb the tree, from crutch to crutch. On each \code{Node}, the
\code{Climb} finds the first child having attribute value equal to the the provided argument.
}
\examples{
data(acme)

#the following are all equivalent
Climb(acme, 'IT', 'Outsource')
Climb(acme, name = 'IT', name = 'Outsource')
Climb(acme, 'IT')$Climb('Outsource')
Navigate(acme, path = "IT/Outsource")

Climb(acme, name = 'IT')

Climb(acme, position = c(2, 1))
#or, equivalent:
Climb(acme, position = 2, position = 1)
Climb(acme, name = "IT", cost = 250000)

tree <- CreateRegularTree(5, 2)
tree$Climb(c("1", "1"), position = c(2, 2))$path

}
\seealso{
\code{\link{Node}}

\code{\link{Navigate}}
}
