% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{Clone}
\alias{Clone}
\title{Clones a tree (creates a deep copy)}
\usage{
Clone(node)
}
\arguments{
\item{node}{the root node of the tree or sub-tree to clone}
}
\value{
the clone of the tree
}
\description{
Clones a tree (creates a deep copy)
}
\examples{
data(acme)
}

