% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.tree.R
\docType{package}
\name{data.tree}
\alias{data.tree}
\alias{data.tree-package}
\title{data.tree hierarchical data for R}
\description{
Hierarchical data is ubiquitous. However, no general-use tree data structure is available in R,
similar to, say, data.frame for tabular data. This package allows building trees of hierarchical
data, traverse the trees and run operations on nodes or collect data. The package also contains
utility functions to sort, to print, and to construct a tree from a data.frame and vice versa.
}

