% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/node.R
\name{Find}
\alias{Find}
\title{Find a \code{Node} by its path}
\usage{
Find(...)
}
\arguments{
\item{...}{the names of the nodes in the path}
}
\value{
the \code{Node} having path \code{...}, or \code{NULL} if such a path does not exist
}
\description{
Find returns the \code{Node} at path \code{...}. The path is relative to the \code{Node} on which this method is called. Each argument provided corresponds to an
element in the path, specified by the \code{Node}'s name.
}
\examples{
data(acme)
acme$Find('IT', 'Outsource')$name
#This is equivalent to:
acme$Find('IT')$Find('Outsource')$name
acme$Find('X', 'Y', 'Z')
}
\seealso{
\code{\link{Node}}
}
\keyword{internal}

