% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_ape.R
\name{GetPhyloNr}
\alias{GetPhyloNr}
\title{Gets the node nr in phylo context}
\usage{
GetPhyloNr(x, type = c("node", "edge"))
}
\arguments{
\item{x}{The Node}

\item{type}{Either "node" (the default) or "edge"}
}
\value{
an integer representing the node
}
\description{
Use this function when plotting a Node as a phylo, to set custom
labels to plot.
}
\examples{
library(ape)
library(data.tree)
data(acme)
ap <- as.phylo(acme)
plot(ap)
nodelabels("IT Dep.", GetPhyloNr(acme$Climb("IT")))
edgelabels("Good!", GetPhyloNr(acme$Climb("IT", "Switch to R"), "edge"))
}

