% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_actives.R
\name{isNotLeaf}
\alias{isNotLeaf}
\title{Check if a \code{Node} is not a leaf}
\usage{
isNotLeaf(node)
}
\arguments{
\item{node}{The Node to test.}
}
\value{
FALSE if the Node is a leaf, TRUE otherwise
}
\description{
Check if a \code{Node} is not a leaf
}
