% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_actives.R
\name{isNotRoot}
\alias{isNotRoot}
\title{Check if a \code{Node} is not a root}
\usage{
isNotRoot(node)
}
\arguments{
\item{node}{The Node to test.}
}
\value{
FALSE if the Node is the root, TRUE otherwise
}
\description{
Check if a \code{Node} is not a root
}

