% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{mushroom}
\alias{mushroom}
\title{Sample Data: Data Used by the ID3 Vignette}
\format{
data.frame
}
\usage{
data(mushroom)
}
\description{
mushroom contains attributes of mushrooms. We can use this data to predict a
mushroom's toxicity based on its attributes.
The attributes available in the data set are:
}
\details{
\itemize{
  \item color the color of a mushroom
  \item size whether a mushroom is small or large
  \item points whether a mushroom has points
  \item edibility whether a mushroom is edible or toxic
}
}
\keyword{datasets}
