% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{Node}
\alias{Node}
\title{Node}
\format{An \code{\link{R6Class}} generator object}
\usage{
Node
}
\description{
A generic node in a tree
}
\section{Fields}{

\describe{
\item{\code{children}}{A list of children}

\item{\code{parent}}{The node's parent Node}
}}
\section{Methods}{


\describe{
  \item{\code{AddChild(name)}}{Creates a new \code{Node} called \code{name} and adds it to this \code{Node}.}
  \item{\code{\link{Find}(...)}}{Find a node with path \code{...}}
  \item{\code{\link{Get}(attribute, ..., traversal = "pre-order", assign = NULL, format = NULL)}}{Traverses the tree and collects values along the way.}
  \item{\code{\link{Set}(..., traversal = "pre-order", returnValues = FALSE)}}{Traverses the tree and assigns attributes along the way.}
  \item{\code{\link{Aggregate}(attribute, fun, ...)}}{Traverses the tree and calls \code{fun(children$Aggregate(...))} on each node. }
  \item{\code{\link{Sort}(attribute, ..., decreasing = FALSE, recursive = TRUE)}}{Sorts the children of a node according to \code{attribute}}
  \item{\code{\link{ToDataFrame}(row.names = NULL, optional = FALSE, ...)}}{Converts the tree below this \code{Node} to a \code{data.frame}}
}
}
\keyword{datasets}

