% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{Node}
\alias{Node}
\title{Node}
\format{An \code{\link{R6Class}} generator object}
\usage{
Node$new()
}
\description{
A generic node in a tree
}
\section{Fields}{

\describe{
\item{\code{children}}{A list of children}

\item{\code{parent}}{The node's parent Node}
}}
\section{Methods}{

\describe{
  \item{\code{AddChild(name = "MyNode")}}{Creates a new Node called \code{name} and adds it to this node.}
}
}
\keyword{datasets}

