% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/node.R
\docType{class}
\name{Node}
\alias{Node}
\title{Node}
\format{An \code{\link{R6Class}} generator object}
\usage{
Node
}
\description{
A generic node in a tree
}
\section{Fields}{

\describe{
\item{\code{children}}{A list of children}

\item{\code{parent}}{The node's parent Node}
}}
\section{Methods}{


\describe{
  \item{\code{AddChild(name)}}{Creates a new \code{Node} called \code{name} and adds it to this \code{Node}.}
  \item{\code{\link{Get}(attribute, ..., traversal = "pre-order", assign = NULL, format = NULL)}}{Traverse the tree and collect values along the way.}
}
}
\keyword{datasets}

