% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node.R
\docType{data}
\name{NODE_RESERVED_NAMES_CONST}
\alias{NODE_RESERVED_NAMES_CONST}
\title{Names that are reserved by the Node class.}
\format{\preformatted{ chr [1:37] "AddChild" "AddChildNode" "AddSibling" "AddSiblingNode" ...
}}
\usage{
NODE_RESERVED_NAMES_CONST
}
\description{
These are reserved by the Node class, you cannot use these as
attribute names.
Note also that all fields starting with a . are reserved.
}
\keyword{datasets}

