% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node.R
\docType{data}
\name{NODE_RESERVED_NAMES_CONST}
\alias{NODE_RESERVED_NAMES_CONST}
\title{Names that are reserved by the Node class.}
\format{An object of class \code{character} of length 40.}
\usage{
NODE_RESERVED_NAMES_CONST
}
\description{
These are reserved by the Node class, you cannot use these as 
attribute names.
Note also that all fields starting with a . are reserved.
}
\keyword{datasets}
