% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/util.R
\name{print.Node}
\alias{print.Node}
\title{Print a Node}
\usage{
\method{print}{Node}(x, ...)
}
\arguments{
\item{x}{The Node}

\item{...}{Additional parameters to be printed as columns, as a string.}
}
\description{
Print a Node in a human-readable fashion.
}
\details{
Print the Node in a human-readable fashion.
}
\examples{
data(acme)
print(acme, "cost", "p")
}

