% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_doc.R, R/node_methods.R
\name{Prune}
\alias{Prune}
\title{Prunes a tree by applying the filterFun to each node.}
\usage{
Prune(node, pruneFun)

Prune(node, pruneFun)
}
\arguments{
\item{pruneFun}{a function taking a \link{\code{Node}} as an argument, and returning TRUE if the Node
and its descendants should be kept, FALSE otherwise.}

\item{traversal}{one of the traversal modes (pre-order, post-order)}

\item{filterFun}{A function taking a \code{\link{Node}} as an argument, an returning \code{TRUE}
if the \code{\link{Node}} should be kept, false otherwise.}

\item{The}{node whose children should be pruned}
}
\value{
A Node
}
\description{
You can apply this to any \code{\link{Node}}
in a tree; the filterFun is applied to the children of the \code{\link{Node}} only. Note that as
soon as the filter criteria is not matched anymore, the algorithm stops, and neither the \code{\link{Node}}
nor its children are kept, regardless of whether the child would satisfy the filter.

Prunes a tree
}
\keyword{internal}

