% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods_sideeffect.R
\name{Prune}
\alias{Prune}
\title{Prunes a tree}
\usage{
Prune(node, pruneFun)
}
\arguments{
\item{pruneFun}{a function taking a \code{\link{Node}} as an argument, and returning TRUE if the Node
and its descendants should be kept, FALSE otherwise.}

\item{The}{node whose children should be pruned}
}
\value{
the number of nodes removed
}
\description{
Prunes a tree
}
\keyword{internal}

