% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_methods_sideeffect.R
\name{Revert}
\alias{Revert}
\title{Reverts the sort order of a \code{Node}'s children.}
\usage{
Revert(node, recursive = TRUE)
}
\arguments{
\item{node}{the Node whose children's sort order is to be reverted}

\item{recursive}{If \code{TRUE}, then revert is called recursively on
all children.}
}
\value{
returns the Nodel invisibly (for chaining)
}
\description{
Reverts the sort order of a \code{Node}'s children.
}
\seealso{
\code{\link{Node}}

\code{\link{Sort}}
}
