% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R
\name{Sort}
\alias{Sort}
\title{Sort Children of a Node or an Entire Tree}
\usage{
Sort(node, attribute, ..., decreasing = FALSE, recursive = TRUE)
}
\arguments{
\item{node}{The node whose children are to be sorted}

\item{attribute}{a field, method or function. The result of the attribute determines the
sorting. If it is a function, #' the attribute must take a \code{Node} as a first argument.}

\item{...}{any parameters to be passed on the the attribute (in case it's a method or a
function)}

\item{decreasing}{sort order}

\item{recursive}{if \code{TRUE}, Sort will be called recursively on the \code{Node}'s children.
This allows sorting an entire tree.}
}
\value{
Returns the node on which Sort is called, invisibly. This can be useful to chain Node methods.
}
\description{
You can sort with respect to any argument of the tree.
}
\examples{
data(acme)
acme$Do(function(x) x$totalCost <- Aggregate(x, "cost", sum), traversal = "post-order")
Sort(acme, "totalCost", decreasing = FALSE)
print(acme, "totalCost")
}
\seealso{
\code{\link{Node}}
}

