% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion_dataframe.R
\name{ToDataFrameTree}
\alias{ToDataFrameTree}
\title{Convert a \code{\link{Node}} to a \code{data.frame}}
\usage{
ToDataFrameTree(node, ..., pruneFun = NULL, inheritFromAncestors = FALSE)
}
\arguments{
\item{node}{The node to convert}

\item{...}{the attributes to be added as columns of the data.frame. There are various
options:
\itemize{
 \item a string corresponding to the name of a node attribute
 \item the result of the \code{Node$Get} method
}
If a specific Node does not contain the attribute, the data.frame will contain NA.}

\item{pruneFun}{a function pruning the nodes to be converted}

\item{inheritFromAncestors}{If TRUE, then field-attributes from ancestors are inherited}
}
\description{
Convert a \code{\link{Node}} to a \code{data.frame}
}
\seealso{
\code{\link{Node}}, \code{\link{as.data.frame.Node}}
}

