% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R, R/node_plot.R, R/util.R
\name{SetNodeStyle}
\alias{Draw}
\alias{GetDefaultTooltip}
\alias{SetEdgeStyle}
\alias{SetGraphStyle}
\alias{SetNodeStyle}
\alias{ToGraphViz}
\title{Get a graphviz dot representation of the tree}
\usage{
SetNodeStyle(node, inherit = TRUE, ...)

SetEdgeStyle(node, inherit = TRUE, ...)

SetGraphStyle(root, ...)

Draw(root, direction = c("climb", "descend"), pruneFun = NULL,
  engine = "dot")

ToGraphViz(root, direction = c("climb", "descend"), pruneFun = NULL)

GetDefaultTooltip(node)
}
\arguments{
\item{node}{The \code{\link{Node}} of the data.tree structure on which you would like to set style attributes.}

\item{inherit}{If TRUE, then children will inherit this node's style. Otherwise they inherit from this
node's parent.}

\item{root}{The root \code{\link{Node}} of the data.tree structure to visualize.}

\item{pruneFun}{a function taking a \code{\link{Node}} as an argument, and returning TRUE if the Node
and its descendants should be kept, FALSE otherwise.}
}
\description{
Get a graphviz dot representation of the tree
}
\examples{
data(acme)
SetGraphStyle(acme, rankdir = "TB")
SetEdgeStyle(acme, arrowhead = "vee", color = "grey35", penwidth = 2)
#per default, Node style attributes will be inherited:
SetNodeStyle(acme, style = "filled,rounded", shape = "box", fillcolor = "GreenYellow", fontname = "helvetica", tooltip = GetDefaultTooltip)
SetNodeStyle(acme$IT, fillcolor = "LightBlue", penwidth = "5px")
#inheritance can be avoided:
SetNodeStyle(acme$Accounting, inherit = FALSE, fillcolor = "Thistle", fontcolor = "Firebrick", tooltip = "This is the accounting department")
#use Do to set style on specific nodes:
Do(acme$leaves, function(node) SetNodeStyle(node, shape = "egg"))
Draw(acme)
}

