% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_methods.R, R/node_plot.R, R/util.R
\name{SetNodeStyle}
\alias{GetDefaultTooltip}
\alias{SetEdgeStyle}
\alias{SetNodeStyle}
\alias{ToGraphViz}
\title{Get a graphviz dot representation of the tree}
\usage{
SetNodeStyle(node, inherit = TRUE, ...)

SetEdgeStyle(node, inherit = TRUE, ...)

ToGraphViz(node, direction = c("climb", "descend"),
  graph_attrs = c("rankdir = BT", "layout = dot"), pruneFun = NULL)

GetDefaultTooltip(node)
}
\arguments{
\item{node}{The root \code{\link{Node}} of the data.tree structure to visualize.}

\item{inherit}{If TRUE, then children will inherit this node's style. Otherwise they inherit from this
node's parent.}

\item{pruneFun}{a function taking a \code{\link{Node}} as an argument, and returning TRUE if the Node
and its descendants should be kept, FALSE otherwise.}

\item{node}{The node}
}
\description{
Get a graphviz dot representation of the tree
}
\examples{
data(acme)
SetNodeStyle(acme, style = "filled,rounded", shape = "box", fillcolor = "GreenYellow", fontname = "helvetica", tooltip = GetDefaultTooltip)
SetEdgeStyle(acme, arrowhead = "vee", color = "grey35", penwidth = 2)
SetNodeStyle(acme$IT, fillcolor = "LightBlue", penwidth = "5px")
SetNodeStyle(acme$Accounting, inherit = FALSE, shape = "circle", fillcolor = "Thistle", fontcolor = "Firebrick", fontsize = 18, width = 2)
render_graph(ToGraphViz(acme))
}

