% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{ToList}
\alias{ToList}
\title{This is an alias for \code{\link{as.list.Node}}.}
\usage{
ToList(x, mode = c("simple", "explicit"), unname = FALSE,
  nameName = ifelse(unname, "name", ""), childrenName = "children",
  rootName = "", ...)
}
\description{
This is an alias for \code{\link{as.list.Node}}.
}

