% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{ToNewick}
\alias{ToNewick}
\title{Write a data.tree to Newick notation}
\usage{
ToNewick(node, heightAttribute = Height, ...)
}
\arguments{
\item{node}{The node to convert}

\item{heightAttribute}{The attribute (field name or function) storing or calculating the height}

\item{...}{parameters that will be passed on the the heightAttributeName, in case it is a function}
}
\description{
Write a data.tree to Newick notation
}
\examples{
data(acme)
acme$ToNewick()
}

