% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/node_conversion.R
\name{ToNewick}
\alias{ToNewick}
\title{Write a data.tree to Newick notation}
\usage{
ToNewick(node, heightAttributeName = "Height", ...)
}
\arguments{
\item{node}{The node to convert}

\item{heightAttributeName}{The name of the attribute or function storing the height}

\item{...}{parameters that will be passed on the the heightAttributeName, in case it is a function}
}
\description{
Write a data.tree to Newick notation
}

